requirements.txt
audioread>=2.1.8
bokeh==2.0.2
certifi>=2019.11.28
cffi>=1.14.0,<=1.14.6
chardet>=3.0.4
click==7.1.1
cloudpickle>=1.3.0
cycler==0.10.0
dask>=2.15.0,<=2.30.0
decorator>=4.4.2,<=5.1.0
dill==0.3.1.1
distributed>=2.15.0,<=2.30.1
fastparquet==0.3.3
fsspec==0.7.3
HeapDict==1.0.1
idna>=2.8,<2.9
imageio==2.8.0
Jinja2>=2.11.3
joblib==0.14.1
kiwisolver==1.2.0
librosa>=0.7.2
llvmlite==0.36.0
locket==0.2.0
MarkupSafe>=2.0
matplotlib==3.2.1
msgpack==1.0.0
networkx==2.4
numba>=0.52.0
numpy>=1.18.1,<=1.19.5
packaging==20.3
pandas==1.0.3
partd==1.1.0
pickle-mixin==1.0.2
Pillow==7.1.1
pony==0.7.13
psutil==5.7.0
pyarrow>=3.0.0
psycopg2==2.8.6
pycparser==2.20
pygraphviz==1.5
pymongo==3.10.1
pyparsing>=2.4.7,<=3.0.0a1
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
requests>=2.22.0
resampy==0.2.2
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Shapely==1.7.0
six>=1.14.0
sortedcontainers==2.1.0
SoundFile==0.10.3.post1
tblib==1.6.0
thrift==0.13.0
toolz==0.10.0
tornado>=6.0.4
tqdm==4.45.0
typing-extensions>=3.7.4.2
urllib3>=1.25.7
zict==2.0.0
